Update mcmod.info
Just added the lacking information so that the mod gets listed as Tesla Arsenal instead of "Example mod" in MultiMC.
The only other thing that needs to be done, that only you can do, is to edit the credits and decide upon what author name you want (koxhack or knoxhack). You might also want to just check over the rest, for example 'dependencies'.